ASoC: wcd9xxx: Disconnect slimbus channel manually
Depending on the ADSP's operating mode, it's possible that slimbus
channel removal request is NACKed.  Since current codec driver has
an assumption that all channel removal requests are guaranteed to be
delivered, when it's NACKed codec would have overflow error.
Fix this overflow issue by waiting channel removal IRQ when requests
channel removal then close the channel manually if there was no removal
IRQ as this likely means channel removal request was NACked.

CRs-fixed: 417075

+++ include/linux/mfd/wcd9xxx/core.h
@@ -131,7 +131,7 @@ struct wcd9xxx_codec_dai_data {
 	u32 bit_width;				/* sit width 16,24,32   */
 	struct list_head wcd9xxx_ch_list;	/* channel list         */
 	u16 grph;				/* slimbus group handle */
-	u32 ch_mask;
+	unsigned long ch_mask;
 	wait_queue_head_t dai_wait;
 };
